+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.23.0] - 2020-12-14
+
+This version encompasses work for a POC of recreating GitHub's 'Discussions' feature, and depends on [v0.7.0 Backend](https://github.com/thiskevinwang/rds-ts-node-server/compare/v0.6.3...v0.7.0)
+
+### Added
+
+- `draft.js`
+- `secondary` prop & styles to `SubmitButton`
+- `RichEditor` component
+- extracted `components/App/Shared/TextField`
+- `yup` for formik validation
+
+### Changed
+
+- enabled a `fs.Fieldset` `hoverable` prop
+
+### Fixed
+
+- `useWindowScrollPercent` sometimes causing `Blob`'s path to throw a NaNNaNNaN error
+
 ## [v0.22.1] - 2020-12-09
 
 I _think_ this finally fixes exif-data-caused orientation problems as well as incorrect cropping with `react-crop`.
@@ -438,6 +458,7 @@ New:
 
 ### Added
 
+[v0.23.0]: https://github.com/thiskevinwang/coffee-code-climb/compare/v0.22.1...v0.23.0
 [v0.22.1]: https://github.com/thiskevinwang/coffee-code-climb/compare/v0.22.0...v0.22.1
 [v0.22.0]: https://github.com/thiskevinwang/coffee-code-climb/compare/v0.21.2...v0.22.0
 [v0.21.2]: https://github.com/thiskevinwang/coffee-code-climb/compare/v0.21.1...v0.21.2